inherited common methods in components from abstract class
