+++ src/components/board.js
@@ -1,4 +1,4 @@
-import AbstractComponent from "./avstract-component";
+import AbstractComponent from "./abstract-component";
 
 const createBoardTemplate = () => {
   return (
@@ -7,7 +7,7 @@ const createBoardTemplate = () => {
 };
 
 
-export default class Board extends AbstractComponent{
+export default class Board extends AbstractComponent {
   getTemplate() {
     return createBoardTemplate();
   }
