+++ src/components/filter.js
@@ -1,4 +1,4 @@
-import AbstractComponent from "./avstract-component";
+import AbstractComponent from "./abstract-component";
 
 const createFilterMarkup = (filter, isChecked) => {
   const {title, count} = filter;
