+++ src/components/no-tasks.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import AbstractComponent from "./abstract-component";
 
 const createNoTasksTemplate = () => {
   return (
@@ -9,24 +9,8 @@ const createNoTasksTemplate = () => {
 };
 
 
-export default class NoTasks {
-  constructor() {
-    this._element = null;
-  }
-
+export default class NoTasks extends AbstractComponent {
   getTemplate() {
     return createNoTasksTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
