+++ src/components/task-form.js
@@ -1,5 +1,6 @@
 import {colorsMockData, daysMockData, monthNames} from '../const.js';
-import {formatTime, createElement} from '../utils.js';
+import {formatTime} from '../utils/common.js';
+import AbstractComponent from "./abstract-component";
 
 const createColorsMarkup = (colors, currentColor) => {
   return colors
@@ -96,7 +97,7 @@ const createTaskFormTemplate = (task) => {
                 <use xlink:href="#wave"></use>
               </svg>
             </div>
-  
+
             <div class="card__textarea-wrap">
               <label>
                 <textarea
@@ -106,14 +107,14 @@ const createTaskFormTemplate = (task) => {
                 >${description}</textarea>
               </label>
             </div>
-  
+
             <div class="card__settings">
               <div class="card__details">
                 <div class="card__dates">
                   <button class="card__date-deadline-toggle" type="button">
                     date: <span class="card__date-status">${isDateShowing ? `yes` : `no`}</span>
                   </button>
-  
+
                   ${isDateShowing ?
       `<fieldset class="card__date-deadline">
                         <label class="card__input-deadline-wrap">
@@ -126,11 +127,11 @@ const createTaskFormTemplate = (task) => {
                           />
                         </label>
                     </fieldset>` : ``}
-  
+
                   <button class="card__repeat-toggle" type="button">
                     repeat:<span class="card__repeat-status">${isRepeatingTask ? `yes` : `no`}</span>
                   </button>
-  
+
                   ${isRepeatingTask ?
       `<fieldset class="card__repeat-days">
                       <div class="card__repeat-days-inner">
@@ -138,12 +139,12 @@ const createTaskFormTemplate = (task) => {
                       </div>
                     </fieldset>` : ``}
                 </div>
-  
+
                 <div class="card__hashtag">
                   <div class="card__hashtag-list">
                     ${tagsMarkup}
                   </div>
-  
+
                   <label>
                     <input
                       type="text"
@@ -154,7 +155,7 @@ const createTaskFormTemplate = (task) => {
                   </label>
                 </div>
               </div>
-  
+
               <div class="card__colors-inner">
                 <h3 class="card__colors-title">Color</h3>
                 <div class="card__colors-wrap">
@@ -162,7 +163,7 @@ const createTaskFormTemplate = (task) => {
                 </div>
               </div>
             </div>
-  
+
             <div class="card__status-btns">
               <button class="card__save" type="submit">save</button>
               <button class="card__delete" type="button">delete</button>
@@ -173,25 +174,14 @@ const createTaskFormTemplate = (task) => {
   );
 };
 
-export default class TaskForm {
+export default class TaskForm extends AbstractComponent {
   constructor(task) {
+    super();
+
     this._task = task;
-    this._element = null;
   }
 
   getTemplate() {
     return createTaskFormTemplate(this._task);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
