+++ src/components/task.js
@@ -1,5 +1,6 @@
 import {monthNames} from '../const.js';
-import {formatTime, createElement} from '../utils.js';
+import {formatTime} from '../utils/common.js';
+import AbstractComponent from "./abstract-component";
 
 const createHashtagsMarkup = (hashtags) => {
   return hashtags
@@ -79,26 +80,14 @@ const createTaskTemplate = (task) => {
   );
 };
 
-export default class Task {
+export default class Task extends AbstractComponent {
   constructor(task) {
-    this._task = task;
+    super();
 
-    this._element = null;
+    this._task = task;
   }
 
   getTemplate() {
     return createTaskTemplate(this._task);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
