+++ src/utils/common.js
@@ -1,8 +1,3 @@
-const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
 const getRandomInRange = function (min, max) {
   return Math.floor(Math.random() * (max - min + 1)) + min;
 };
@@ -37,25 +32,4 @@ const formatTime = (date) => {
   return `${hours}:${minutes} ${interval}`;
 };
 
-const renderComponent = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-export {getRandomInRange, getRandomArrayItem, getRandomDate, renderComponent, RenderPosition,
-  formatTime, createElement
-};
-
+export {getRandomInRange, getRandomArrayItem, getRandomDate, formatTime};
